==> bin/3_1_segment_qc.R <==
#!/usr/bin/env Rscript
args = commandArgs(trailingOnly=TRUE)
pathBase = args[1]
minSegmentReads = as.integer(args[2])
percentTrimmed = as.integer(args[3])
percentStitched = as.integer(args[4])
percentAligned = as.integer(args[5])
percentSaturation = as.integer(args[6])
minNegativeCount = as.integer(args[7])
maxNTCCount = as.integer(args[8])
minNuclei = as.integer(args[9])
minArea = as.integer(args[10])
region1 = args[11]


load(sprintf('%s/image/2_sample_overview.RData', pathBase))

############################
###   3.1 - Segment QC   ###
############################

#####################################
###   3.1.1 - Select Segment QC   ###
#####################################

library(GeomxTools)

# Shift counts to one
data <- shiftCountsOne(data, useDALogic = TRUE)

# Default QC cutoffs are commented in () adjacent to the respective parameters
# study-specific values were selected after visualizing the QC results in more
# detail below
QC_params <-
  list(minSegmentReads = minSegmentReads, # Minimum number of reads (1000)
       percentTrimmed = percentTrimmed,    # Minimum % of reads trimmed (80%)
       percentStitched = percentStitched,   # Minimum % of reads stitched (80%)
       percentAligned = percentAligned,    # Minimum % of reads aligned (80%)
       percentSaturation = percentSaturation, # Minimum sequencing saturation (50%)
       minNegativeCount = minNegativeCount,   # Minimum negative control counts (10)
       maxNTCCount = maxNTCCount,     # Maximum counts observed in NTC well (1000)
       minNuclei = minNuclei,         # Minimum # of nuclei estimated (100)
       minArea = minArea)         # Minimum segment area (5000)
data <- setSegmentQCFlags(data, qcCutoffs = QC_params)        

# Collate QC Results
QCResults <- protocolData(data)[["QCFlags"]]
flag_columns <- colnames(QCResults)
QC_Summary <- data.frame(Pass = colSums(!QCResults[, flag_columns]),
                         Warning = colSums(QCResults[, flag_columns]))
QCResults$QCStatus <- apply(QCResults, 1L, function(x) {
  ifelse(sum(x) == 0L, "PASS", "WARNING")
})
QC_Summary["TOTAL FLAGS", ] <-
  c(sum(QCResults[, "QCStatus"] == "PASS"),
    sum(QCResults[, "QCStatus"] == "WARNING"))

################################################
###   Section 3.1.2 - Visualize Segment QC   ###
################################################

library(ggplot2)
library(knitr)

# Graphical summaries of QC statistics plot function
QC_histogram <- function(assay_data = NULL,
                         annotation = NULL,
                         fill_by = NULL,
                         thr = NULL,
                         scale_trans = NULL) {
  plt <- ggplot(assay_data,
                aes_string(x = paste0("unlist(`", annotation, "`)"),
                           fill = fill_by)) +
    geom_histogram(bins = 50) +
    geom_vline(xintercept = thr, lty = "dashed", color = "black") +
    theme_bw() + guides(fill = "none") +
    facet_wrap(as.formula(paste("~", fill_by)), nrow = 4) +
    labs(x = annotation, y = "Segments, #", title = annotation)
  if(!is.null(scale_trans)) {
    plt <- plt +
      scale_x_continuous(trans = scale_trans)
  }
  plt
}

# Column to plot
col_by <- "segment"

QC_histogram(sData(data), "Trimmed (%)", col_by, 80)
ggsave(sprintf("%s/plots/3_1_2_trimmed_percentage.png", pathBase), device='png')

QC_histogram(sData(data), "Stitched (%)", col_by, 80)
ggsave(sprintf("%s/plots/3_1_2_stitched_percentage.png", pathBase), device='png')

QC_histogram(sData(data), "Aligned (%)", col_by, 75)
ggsave(sprintf("%s/plots/3_1_2_aligned_percentage.png", pathBase), device='png')

QC_histogram(sData(data), "Saturated (%)", col_by, 50) +
  labs(title = "Sequencing Saturation (%)", x = "Sequencing Saturation (%)")
ggsave(sprintf("%s/plots/3_1_2_saturation_percentage.png", pathBase), device='png')

QC_histogram(sData(data), "area", col_by, 1000, scale_trans = "log10")
ggsave(sprintf("%s/plots/3_1_2_area.png", pathBase), device='png')

QC_histogram(sData(data), "nuclei", col_by, 20)
ggsave(sprintf("%s/plots/3_1_2_nuclei.png", pathBase), device='png')


# calculate the negative geometric means for each module
negativeGeoMeans <- 
  esBy(negativeControlSubset(data), 
       GROUP = "Module", 
       FUN = function(x) { 
         assayDataApply(x, MARGIN = 2, FUN = ngeoMean, elt = "exprs") 
       }) 
protocolData(data)[["NegGeoMean"]] <- negativeGeoMeans

# explicitly copy the Negative geoMeans from sData to pData
negCols <- paste0("NegGeoMean_", modules)
pData(data)[, negCols] <- sData(data)[["NegGeoMean"]]
for(ann in negCols) {
  plt <- QC_histogram(pData(data), ann, col_by, 2, scale_trans = "log10")
  print(plt)
  ggsave(sprintf("%s/plots/3_1_2_NegGeoMean_%s.png", pathBase, ann), device='png')
}

# detatch neg_geomean columns ahead of aggregateCounts call
pData(data) <- pData(data)[, !colnames(pData(data)) %in% negCols]

# show all NTC values
ntc_count_table <- data.frame(table(NTC_Count = sData(data)$NTC))
colnames(ntc_count_table) <- c('NTC Count', '# of Segments')
write.csv(
  ntc_count_table,
  sprintf("%s/data/3_1_2_table_ntc_count.csv", pathBase),
  row.names=FALSE
)

QC_Summary <- data.frame(QC_Summary)
write.csv(
  QC_Summary,
  sprintf("%s/data/3_1_2_table_qc_summary.csv", pathBase)
)


###################################################
###   Section 3.1.3 - Remove flagged segments   ###
###################################################

data <- data[, QCResults$QCStatus == "PASS"]
write.csv(
  data.frame(dim(data)),
  sprintf("%s/data/3_1_3_dimensions_after_segment_qc.csv", pathBase)
)

# Save image
save.image(sprintf('%s/image/3_1_segment_qc.RData', pathBase))
